+++ api/src/Application/Actions/Occasion/CreateParticipantOccasionAction.php
@@ -6,6 +6,7 @@ namespace App\Application\Actions\Occasion;
 
 use App\Application\Actions\Action;
 use App\Application\Serializable\Utilisateur\SerializableUtilisateur;
+use App\Application\Service\MailerService;
 use App\Domain\Occasion\OccasionRepository;
 use App\Domain\Utilisateur\UtilisateurRepository;
 use Psr\Http\Message\ResponseInterface as Response;
@@ -16,15 +17,18 @@ class CreateParticipantOccasionAction extends Action
 {
   protected $occasionRepository;
   protected $utilisateurRepository;
+  private $mailerService;
 
   public function __construct(
     LoggerInterface $logger,
     OccasionRepository $occasionRepository,
-    UtilisateurRepository $utilisateurRepository
+    UtilisateurRepository $utilisateurRepository,
+    MailerService $mailerService
   ) {
     parent::__construct($logger);
     $this->occasionRepository = $occasionRepository;
     $this->utilisateurRepository = $utilisateurRepository;
+    $this->mailerService = $mailerService;
   }
 
   protected function action(): Response
@@ -40,6 +44,9 @@ class CreateParticipantOccasionAction extends Action
     $occasion->addParticipant($participant);
 
     $occasion = $this->occasionRepository->update($occasion);
+
+    $this->mailerService->envoieMailAjoutParticipant($this->request, $participant, $occasion);
+
     return $this->respondWithData(new SerializableUtilisateur($participant));
   }
 }