+++ api/src/Application/Actions/Utilisateur/CreateUtilisateurReinitMdpAction.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Application\Actions\Utilisateur;
 
 use App\Application\Serializable\Utilisateur\SerializableUtilisateur;
+use App\Application\Service\MailerService;
 use App\Application\Service\PasswordService;
 use App\Domain\Utilisateur\UtilisateurRepository;
 use Psr\Http\Message\ResponseInterface as Response;
@@ -13,15 +14,18 @@ use Psr\Log\LoggerInterface;
 class CreateUtilisateurReinitMdpAction extends OneUtilisateurAction
 {
   private $passwordService;
+  private $mailerService;
 
   public function __construct(
     LoggerInterface $logger,
     UtilisateurRepository $utilisateurRepository,
-    PasswordService $passwordService
+    PasswordService $passwordService,
+    MailerService $mailerService
   )
   {
     parent::__construct($logger, $utilisateurRepository);
     $this->passwordService = $passwordService;
+    $this->mailerService = $mailerService;
   }
 
   protected function action(): Response
@@ -33,7 +37,9 @@ class CreateUtilisateurReinitMdpAction extends OneUtilisateurAction
     $utilisateur->setMdp(password_hash($mdp, PASSWORD_DEFAULT));
     $this->utilisateurRepository->update($utilisateur);
 
-    return $this->respondWithData(new SerializableUtilisateur($utilisateur, true, $mdp));
+    $this->mailerService->envoieMailReinitialisationMdp($this->request, $utilisateur, $mdp);
+
+    return $this->respondWithData(new SerializableUtilisateur($utilisateur, true));
   }
 
 }