+++ api/src/Application/Actions/Utilisateur/UtilisateurAction.php
@@ -7,6 +7,7 @@ use App\Application\Actions\Action;
 use App\Domain\Utilisateur\UtilisateurRepository;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Log\LoggerInterface;
+use Slim\Exception\HttpBadRequestException;
 
 class UtilisateurAction extends Action
 {
@@ -30,5 +31,18 @@ class UtilisateurAction extends Action
         $this->assertAuth();
 
         return $this->response;
+    }
+
+    /**
+     * Vérifie que l'utilisateur authentifié est admin
+     */
+    protected function assertEstEmail($email, $warning = null)
+    {
+        if (!filter_var($email, FILTER_VALIDATE_EMAIL)) {
+            throw new HttpBadRequestException($this->request, "$email n'est pas un email valide");
+            if (!is_null($warning)) {
+                $this->logger->warning($warning);
+            }
+        }
     }
 }